feat(snippets): add C and C++ snippets for VSCode
This commit introduces a new set of snippets for C and C++ languages
to be used in VSCode. The snippets include common code structures such
as loops, conditionals, file operations, and more. This will enhance
the coding experience by reducing the amount of boilerplate code that
needs to be written.